+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Casing of get *O*ne ME-model.
 - Deduplicate tool selection on our side.
 
+### Removed
+- Semantic router.
+
 ## [v0.7.2] - 30.07.2025
 
 ### Added